[#7] refactor :recycle: Pitch모델 속성과 InningHistory 이름 변경, GameHeader 모델 제거
