+++ iOS/BaseballGame/BaseballGame/Utilities/Network/Mock/MockPitchHistoryUseCase.swift
@@ -1,0 +1,9 @@
+//
+//  MockPitchHistoryUseCase.swift
+//  BaseballGame
+//
+//  Created by delma on 2020/05/14.
+//  Copyright © 2020 delma. All rights reserved.
+//
+
+import Foundation
